Create better separation between packages
Create separate packages for forms & API related logic where it makes
sense. The key point to consider when creating packages is what are the
dependency relations between them. For example, the cyberstorm component
library does not have to depend on the react hook form library, but we
still need to implement the forms somewhere. As such, it makes the most
sense to spliti forms related components to their own package.

This is not a final solution, but a significant improvement to what was
in place prior.
