+++ packages/cyberstorm-forms/README.md
@@ -1,0 +1,8 @@
+# cyberstorm-forms
+
+Cyberstorm form components
+
+## Scripts
+
+- `yarn run build`: Builds the project
+- `yarn run dev`: Builds the project & watches files for changes, triggering a rebuild
