+++ packages/cyberstorm-forms/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@thunderstore/cyberstorm-forms",
+  "version": "0.1.0",
+  "description": "Cyberstorm form components",
+  "repository": "https://github.com/thunderstore-io/thunderstore-ui/",
+  "main": "dist/thunderstore-cyberstorm-forms.cjs.js",
+  "module": "dist/thunderstore-cyberstorm-forms.esm.js",
+  "types": "dist/thunderstore-cyberstorm-forms.cjs.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc",
+    "dev": "tsc --watch"
+  },
+  "dependencies": {
+    "@thunderstore/cyberstorm": "*",
+    "@types/react": "^18.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.48.2",
+    "zod": "^3.22.2"
+  },
+  "devDependencies": {
+    "typescript-plugin-css-modules": "^3.4.0"
+  }
+}
