+++ packages/cyberstorm-forms/src/components/FormSubmitButton.module.css
@@ -1,0 +1,9 @@
+.spinningIcon {
+  animation: rotate 2s linear infinite;
+}
+
+@keyframes rotate {
+  100% {
+    transform: rotate(360deg);
+  }
+}
