+++ packages/cyberstorm-forms/src/components/FormTextInput.module.css
@@ -1,0 +1,5 @@
+.errorMessage {
+  color: #f1385a;
+  font-weight: 500;
+  font-size: 0.75rem;
+}
