+++ packages/cyberstorm-forms/src/components/FormTextInput.tsx
@@ -1,0 +1,42 @@
+"use client";
+
+import { z, ZodObject } from "zod";
+import { ZodRawShape } from "zod/lib/types";
+import { Path, useController } from "react-hook-form";
+import { TextInput } from "@thunderstore/cyberstorm";
+import styles from "./FormTextInput.module.css";
+
+export type FormTextInputProps<
+  Schema extends ZodObject<Z>,
+  Z extends ZodRawShape
+> = {
+  // The schema is required to allow TS to infer valid values for the name field
+  schema: Schema;
+  name: Path<z.infer<Schema>>;
+  placeholder?: string;
+};
+export function FormTextInput<
+  Schema extends ZodObject<Z>,
+  Z extends ZodRawShape
+>({ name, placeholder }: FormTextInputProps<Schema, Z>) {
+  const {
+    field,
+    fieldState: { isDirty, invalid, error },
+    formState: { isSubmitting, disabled },
+  } = useController({ name });
+
+  return (
+    <>
+      <TextInput
+        {...field}
+        ref={field.ref}
+        placeholder={placeholder}
+        color={isDirty ? (invalid ? "red" : "green") : undefined}
+        disabled={isSubmitting || disabled}
+      />
+      {error && <span className={styles.errorMessage}>{error.message}</span>}
+    </>
+  );
+}
+
+FormTextInput.displayName = "FormTextInput";
