+++ packages/cyberstorm-forms/src/forms/CreateTeamForm.module.css
@@ -25,19 +25,3 @@
   padding-top: var(--space--16);
   border-top: var(--border);
 }
-
-.spinningIcon {
-  animation: rotate 2s linear infinite;
-}
-
-.errorMessage {
-  color: #f1385a;
-  font-weight: 500;
-  font-size: 0.75rem;
-}
-
-@keyframes rotate {
-  100% {
-    transform: rotate(360deg);
-  }
-}
