+++ packages/cyberstorm-forms/src/useFormToaster.ts
@@ -1,0 +1,32 @@
+"use client";
+
+import { useToast } from "@thunderstore/cyberstorm/src/components/Toast/Provider";
+
+export type UseFormToasterArgs = {
+  successMessage: string;
+};
+export type UseFormToasterReturn = {
+  onSubmitSuccess: () => void;
+  onSubmitError: () => void;
+};
+export function useFormToaster({
+  successMessage,
+}: UseFormToasterArgs): UseFormToasterReturn {
+  const toast = useToast();
+
+  return {
+    onSubmitSuccess: () => {
+      toast.addToast({
+        variant: "success",
+        message: successMessage,
+        duration: 30000,
+      });
+    },
+    onSubmitError: () => {
+      toast.addToast({
+        variant: "danger",
+        message: "Unknown error occurred. The error has been logged",
+      });
+    },
+  };
+}
