+++ packages/cyberstorm/package.json
@@ -24,7 +24,6 @@
     "@fortawesome/pro-solid-svg-icons": "6.2.0",
     "@fortawesome/pro-thin-svg-icons": "6.2.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@hookform/resolvers": "^3.3.2",
     "@radix-ui/react-checkbox": "^1.0.1",
     "@radix-ui/react-dialog": "^1.0.2",
     "@radix-ui/react-dropdown-menu": "^2.0.1",
@@ -40,12 +39,10 @@
     "react": "^18.2.0",
     "react-data-table-component": "^7.5.3",
     "react-dom": "^18.2.0",
-    "react-hook-form": "^7.48.2",
     "react-markdown": "^8.0.7",
     "remark-gfm": "^3.0.1",
     "styled-components": "^6.0.0-rc.5",
-    "use-debounce": "^9.0.4",
-    "zod": "^3.22.2"
+    "use-debounce": "^9.0.4"
   },
   "devDependencies": {
     "typescript-plugin-css-modules": "^3.4.0"
