+++ packages/cyberstorm/src/components/Toast/Provider.tsx
@@ -1,3 +1,4 @@
+"use client";
 import * as RadixToast from "@radix-ui/react-toast";
 import {
   createContext,
