+++ packages/thunderstore-api/src/errors.ts
@@ -1,3 +1,7 @@
+export function isApiError(e: Error | ApiError | unknown): e is ApiError {
+  return e instanceof ApiError;
+}
+
 export class ApiError extends Error {
   constructor(message?: string) {
     super(message);
@@ -5,7 +9,7 @@ export class ApiError extends Error {
 
   static createFromResponse(response: Response): ApiError {
     // TODO: Implement response parsing for known error scenarios
-    return new ApiError(response.statusText);
+    return new ApiError(`${response.status}: ${response.statusText}`);
   }
 
   getFieldErrors(): { [key: string]: string[] } {
