+++ packages/ts-api-react-forms/README.md
@@ -1,0 +1,8 @@
+# ts-api-react-forms
+
+React forms logic for the Thunderstore API
+
+## Scripts
+
+- `yarn run build`: Builds the project
+- `yarn run dev`: Builds the project & watches files for changes, triggering a rebuild
