+++ packages/ts-api-react-forms/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@thunderstore/ts-api-react-forms",
+  "version": "0.1.0",
+  "description": "React forms logic for the Thunderstore API",
+  "repository": "https://github.com/thunderstore-io/thunderstore-ui/",
+  "main": "dist/thunderstore-ts-api-react-forms.cjs.js",
+  "module": "dist/thunderstore-ts-api-react-forms.esm.js",
+  "types": "dist/thunderstore-ts-api-react-forms.cjs.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc",
+    "dev": "tsc --watch"
+  },
+  "dependencies": {
+    "@hookform/resolvers": "^3.3.2",
+    "@thunderstore/thunderstore-api": "*",
+    "@types/react": "^18.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.48.2",
+    "zod": "^3.22.2"
+  },
+  "devDependencies": {
+    "typescript-plugin-css-modules": "^3.4.0"
+  }
+}
