+++ packages/ts-api-react/README.md
@@ -1,0 +1,8 @@
+# ts-api-react
+
+Thunderstore API React integration
+
+## Scripts
+
+- `yarn run build`: Builds the project
+- `yarn run dev`: Builds the project & watches files for changes, triggering a rebuild
